flytter Årsak type i egen fil
